proper casting of numpy.floor
numpy.floor on l.23 returns an array of floats, which don't get handled by the %x formatting: "TypeError: %x format: an integer is required, not numpy.float64". A simple cast fixes this.

+++ examples/plotting/server/hover.py
@@ -20,7 +20,7 @@ N = len(x)
 inds = [str(i) for i in np.arange(N)]
 radii = np.random.random(size=N)*0.4 + 1.7
 colors = [
-    "#%02x%02x%02x" % (r, g, 150) for r, g in zip(np.floor(50+2*x), np.floor(30+2*y))
+    "#%02x%02x%02x" % (r, g, 150) for r, g in zip(np.floor(50+2*x).astype(int), np.floor(30+2*y).astype(int))
 ]
 
 foo = list(itertools.permutations("abcdef"))[:N]
